ticket #76 - Starting task from user interface fails and causes corruption
The controller was added the session id to the beginning of the list, but the interface was removing it from the end when calling the proxied function.
This caused the task_key to be truncated and the session_id took its place.

The controller now adds the session_id to the end of the args sent to every function.  This allows it to be removed from the end of the args properly.