Scheduler - Fixing bug with ParallelTask compeltion:
When a job is run on the mainworker this overwrites the WorkerJob for
the original task.  The job has subtask_key set, which
scheduler.send_results detects, and decides that the workerjob is a
subtask.  I was able to fix it by clearing the subtask_key along with
the other attributes keyed on "on_main_worker".

The longer term problem here is that attributes stored within WorkerJob
may be overwritten.  WorkerJob should just remain a mapping of what a
worker is currently doing.  Anything task specific should remain stored
in TaskInstance.  That removes any future issues of attributes being lost.