Tasks - improved get_subtask(...) to support task specific initialization of subtasks
Currently the worker tracks both the instance of the Task, and subtask that are being run.  This works but requires
the worker have knowledge of what type of Task subclass it is.  This design is not extensible.  It is not possible
to create additional task subclasses that have subtasks without altering core code.

improving this logic also allows for subtask starting to be pushed down into Task subclass as well.

Specifics:
   * Task.get_subtask(key) is now an iterative function that calls Task._get_subtask(...)
   * Task._get_subtask() is virtually the same as
   * Task.get_subtask and _get_subtask both take boolean flag clean (default = false) to allow creation of a new subtask instance
   * ParallelTask updated to work with new get_subtask workflow
   * ParallelTask now has set_subtask(...) for setting subtask info.  subtask is lazy loaded when ParallelTask.subtask is accessed or _get_subtask() called