TaskScheduler - updated fetch_task_statuses to work asynchronously
   previously the status_update mechanism would return the current cache and then start an update.
This resulted in the clients of the API receiving a cache that was 3 seconds out of date.  Updating
this to use deferreds allows the requests to wait until the response is ready before returning.  This
reduces the delay to the time it takes to receive current status to the time it takes for all workers to respond

   * first user triggers status fetching, all other users receive a deferred immediately.  The process generates
     a deferred list that callback after all statuses are returned.
   * removed excess iteration and object lookup.  previous implementation iterated workers and then the queue.
     this version only iterates the queue once and requires less lookups by passing dicts to update to matching callbacks